requirements.txt
torch
torchvision
pandas
Pillow
mysql-connector-python
flask
